Refactor some codes
