+++ client/src/pages/DashboardPage.tsx
@@ -1,47 +1,12 @@
 import React, { useState, useEffect } from 'react';
-import { Clock, Calendar, Check } from 'lucide-react';
-import { ClockButton } from '@/components/time/ClockButton';
-import { StatsCard } from '@/components/dashboard/StatsCard';
+import { TimeTrackingCard } from '@/components/dashboard/TimeTrackingCard';
+import { TimeStats } from '@/components/dashboard/TimeStats';
 import { PTOForm } from '@/components/pto/PTOForm';
 import { PTOStatus } from '@/components/pto/PTOStatus';
-import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card';
 import { useAuth } from '@/contexts/AuthContext';
 import { useTime } from '@/contexts/TimeContext';
 import { usePTO } from '@/contexts/PTOContext';
-
-const formatHours = (hours: number): string => {
-  const wholeHours = Math.floor(hours);
-  const minutes = Math.round((hours - wholeHours) * 60);
-  
-  if (wholeHours === 0) {
-    return `${minutes} minutes`;
-  }
-  
-  if (wholeHours === 1) {
-    return `1 hour${minutes > 0 ? ` ${minutes} minutes` : ''}`;
-  }
-  
-  return `${wholeHours} hours${minutes > 0 ? ` ${minutes} minutes` : ''}`;
-};
-
-const formatElapsedTime = (clockInTime: string): string => {
-  const start = new Date(clockInTime).getTime();
-  const now = new Date().getTime();
-  const diffInMinutes = Math.floor((now - start) / (1000 * 60));
-  
-  const hours = Math.floor(diffInMinutes / 60);
-  const minutes = diffInMinutes % 60;
-  
-  if (hours === 0) {
-    return `${minutes} minutes`;
-  }
-  
-  if (hours === 1) {
-    return `1 hour ${minutes} minutes`;
-  }
-  
-  return `${hours} hours ${minutes} minutes`;
-};
+import { formatElapsedTime } from '@/utils/timeFormatters';
 
 const DashboardPage: React.FC = () => {
   const { user } = useAuth();
@@ -66,7 +31,7 @@ const DashboardPage: React.FC = () => {
     return () => {
       document.removeEventListener('visibilitychange', handleVisibilityChange);
     };
-  }, []); // Remove fetchRequests from dependencies since it's stable
+  }, []); 
   
   // Update elapsed time every minute
   useEffect(() => {
@@ -93,49 +58,14 @@ const DashboardPage: React.FC = () => {
       </div>
       
       {/* Stats Row */}
-      <div className="grid grid-cols-1 md:grid-cols-3 gap-4">
-        <StatsCard
-          title="Current Status"
-          value={isClockedIn ? "Clocked In" : "Clocked Out"}
-          icon={<Clock className="h-4 w-4 text-muted-foreground" />}
-          description={isClockedIn ? "You're currently on the clock" : "You're not currently working"}
-          className={isClockedIn ? "border-l-4 border-l-green-500" : "border-l-4 border-l-gray-300"}
-        />
-        
-        <StatsCard
-          title="Hours Today"
-          value={formatHours(totalHoursToday)}
-          icon={<Calendar className="h-4 w-4 text-muted-foreground" />}
-          description="Total hours worked today"
-        />
-        
-        <StatsCard
-          title="Hours This Week"
-          value={formatHours(totalHoursThisWeek)}
-          icon={<Check className="h-4 w-4 text-muted-foreground" />}
-          description="Total hours worked this week"
-        />
-      </div>
+      <TimeStats
+        isClockedIn={isClockedIn}
+        totalHoursToday={totalHoursToday}
+        totalHoursThisWeek={totalHoursThisWeek}
+      />
       
       {/* Clock In/Out Card */}
-      <Card>
-        <CardHeader>
-          <CardTitle>Time Tracking</CardTitle>
-          <CardDescription>
-            Clock in when you start working and clock out when you're done
-          </CardDescription>
-          {currentEntry && (
-            <>
-              <div className="mt-4 text-2xl font-bold">
-                Clocked in since {new Date(currentEntry.clockIn).toLocaleTimeString([], { hour: 'numeric', minute: '2-digit' })}
-              </div>
-            </>
-          )}
-        </CardHeader>
-        <CardContent className="flex justify-center">
-          <ClockButton className="w-full max-w-xs text-lg" />
-        </CardContent>
-      </Card>
+      <TimeTrackingCard currentEntry={currentEntry} />
       
       {/* PTO Status */}
       <PTOStatus />
