+++ client/src/types/index.ts
@@ -1,0 +1,6 @@
+export interface TimeEntry {
+  clockIn: string;
+  clockOut?: string;
+  id: string;
+  userId: string;
+}
